refactor: modularize action handlers
